+++ src/bzflag/EffectsRenderer.h
@@ -46,9 +46,9 @@ public:
   virtual void setRot(const fvec3& rot);
   virtual void setVel(const fvec3& vel);
   virtual void setColor(const fvec4& rgba);
-  virtual void setStartTime(float time);
+  virtual void setStartTime(double time);
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer &) = 0;
 
 protected:
@@ -56,11 +56,11 @@ protected:
   fvec3 rotation;
   fvec3 velocity;
   fvec4 color;
-  float startTime;
-  float lifetime;
-  float lastTime;
-  float deltaTime;
-  float age;
+  double startTime;
+  double lifetime;
+  double lastTime;
+  double deltaTime;
+  double age;
 };
 
 typedef std::vector<BasicEffect*> tvEffectsList;
@@ -71,7 +71,7 @@ public:
   StdSpawnEffect();
   virtual ~StdSpawnEffect();
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 
 protected:
@@ -83,7 +83,7 @@ protected:
 class ConeSpawnEffect : public StdSpawnEffect
 {
 public:
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 };
 
@@ -92,7 +92,7 @@ class RingSpawnEffect : public StdSpawnEffect
 public:
   RingSpawnEffect();
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 
 private:
@@ -109,7 +109,7 @@ public:
   StdShotEffect();
   virtual ~StdShotEffect();
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 
 protected:
@@ -123,7 +123,7 @@ class FlashShotEffect : public StdShotEffect
 public:
   FlashShotEffect();
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 
 private:
@@ -136,7 +136,7 @@ public:
   StdDeathEffect();
   virtual ~StdDeathEffect();
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 
 protected:
@@ -151,7 +151,7 @@ public:
   StdLandEffect();
   virtual ~StdLandEffect();
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 
 protected:
@@ -166,7 +166,7 @@ public:
   StdGMPuffEffect();
   virtual ~StdGMPuffEffect();
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 protected:
   int	  texture;
@@ -180,7 +180,7 @@ public:
   StdRicoEffect();
   virtual ~StdRicoEffect();
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 protected:
   int	  texture;
@@ -194,7 +194,7 @@ public:
   StdShotTeleportEffect(float length, const fvec4* clipPlane);
   virtual ~StdShotTeleportEffect();
 
-  virtual bool update(float time);
+  virtual bool update(double time);
   virtual void draw(const SceneRenderer& sr);
 
 protected:
